dealing with non expression statements and inheritance issues
In the rare cases that we need to extend from a base controller and we
are using the default constructor we still need to properly call super,
that is, we need to make a dummy constructor so that the plugin may add
the proper parameters. We should also keep in mind that not all the
nodes are expressions, some are of the `VariableDeclaration` type, thus
we check first that `fCmd.expression` exists.